+++ change-log.md
@@ -5,6 +5,9 @@
                         
 [//]: #@corifeus-header:end
 
+### v2021.10.195
+Released on 10/26/2021
+* BUGFIX: Electron v15 new window open function update.
 
 ### v2021.10.185
 Released on 10/26/2021
@@ -778,7 +781,7 @@ All my domains ([patrikx3.com](https://patrikx3.com) and [corifeus.com](https://
 
 ---
 
-[**P3X-REDIS-UI**](https://corifeus.com/redis-ui) Build v2021.10.185
+[**P3X-REDIS-UI**](https://corifeus.com/redis-ui) Build v2021.10.195
 
 [![Donate for Corifeus / P3X](https://img.shields.io/badge/Donate-Corifeus-003087.svg)](https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QZVM4V6HVZJW6)  [![Contact Corifeus / P3X](https://img.shields.io/badge/Contact-P3X-ff9900.svg)](https://www.patrikx3.com/en/front/contact) [![Like Corifeus @ Facebook](https://img.shields.io/badge/LIKE-Corifeus-3b5998.svg)](https://www.facebook.com/corifeus.software)